requirements.txt
asgiref==3.8.1
autopep8==2.3.2
babel==2.17.0
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap5==2025.4
cryptography==45.0.2
diff-match-patch==20241021
Django==5.2.1
django-admin-honeypot==1.1.0
django-allauth==65.8.1
django-axes==8.0.0
django-cacheops==7.2
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-countries==7.6.1
dj-database-url==2.1.0
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-formtools==2.5.1
django-health-check==3.18.3
django-import-export==4.3.7
django-js-asset==3.1.2
django-modeltranslation==0.19.14
django-money==3.5.4
django-mptt==0.17.0
django-otp==1.6.0
django-parler==2.3
django-phonenumber-field==8.1.0
django-prometheus==2.3.1
django-redis==5.4.0
django-silk==5.3.2
django-simple-history==3.8.0
django-storages==1.14.6
django-taggit==6.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
funcy==2.0
gprof2dot==2025.4.14
humanize==4.12.3
idna==3.10
jmespath==1.0.1
phonenumbers==9.0.5
pillow==11.2.1
prometheus_client==0.22.0
psutil==7.0.0
psycopg2-binary==2.9.10
py-moneyed==3.0
pycodestyle==2.13.0
pycparser==2.22
PyJWT==2.10.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
qrcode==7.4.2
redis==6.1.0
requests==2.32.3
s3transfer==0.13.0
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.2
urllib3==2.4.0
